backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
dashscope>=1.20.14
requests==2.31.0
python-dotenv==1.0.0